requirements.txt
asgiref==3.7.2
attrs==23.1.0
beautifulsoup4==4.12.2
Django==5.0
django-environ==0.11.2
djangorestframework==3.14.0
drf-spectacular==0.27.0
exceptiongroup==1.2.0
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
model-bakery==1.17.0
packaging==23.2
pluggy==1.3.0
protobuf==4.25.1
psycopg2-binary==2.9.9
pytest==7.4.3
pytest-django==4.7.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
rpds-py==0.15.2
soupsieve==2.5
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
uritemplate==4.1.1
